pyproject.toml
[tool.poetry]
name = "dockenv"
version = "0.1.3"
description = "Simple cli utility for generate docker-compose.yml files"
authors = ["t33m4h"]

[tool.poetry.dependencies]
python = "^3.6"
fire = "^0.2.1"
loguru = "^0.3.2"
pyyaml = "^5.1"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[tool.poetry.scripts]
dockenv = "dockenv.dockenv:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
